pyproject.toml
[tool.poetry]
name = "verimodel"
version = "0.2.0" # Nâng version
description = " AI Supply Chain Firewall  - Scan and detect malicious code in pickle-based ML models."
authors = ["Tran Tuan Anh <Dylan> <dyltran3@gmail.com>"]
license = "MIT"
homepage = "https://github.com/Dylan-Ki/verimodel"
repository = "https://github.com/Dylan-Ki/verimodel"
keywords = ["security", "ai", "ml", "pickle", "model", "firewall", "scanner", "yara", "sandbox"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Security",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

[tool.poetry.dependencies]
python = "^3.10"
typer = {extras = ["all"], version = "^0.9.0"}
rich = "^13.7.0"
docker = "^7.0.0"       # Thư viện mới
yara-python = "^4.5.0"  # Thư viện mới

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "^23.12.0"
flake8 = "^6.1.0"
pip-audit = "^2.7.3" # Thêm để quét CI

[tool.poetry.scripts]
verimodel = "verimodel.cli:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]